calendar date test
