+++ maisel-rampen-booking-big-calendar/src/App.js
@@ -1,17 +1,13 @@
 
 import './App.scss';
 import 'bootstrap/dist/css/bootstrap.min.css';
+import MaiselBaseCalendar from './components/big-calendar/MaiselBaseCalendar'
+import { Calendar as BigCalendar, momentLocalizer, Views as CalendarViews } from 'react-big-calendar';
 
-import { Calendar as BigCalendar, momentLocalizer } from 'react-big-calendar';
-import moment from 'moment';
 import events from './events';
 import { getCurrentDayDate } from './util'
 
 
-const localizer = momentLocalizer(moment);
-console.log(localizer)
-
-//const allViews = Object.keys(BigCalendar.Views).map(k => BigCalendar.Views[k]);
 
 function App() {
   const currentDate = getCurrentDayDate();
@@ -26,16 +22,7 @@ function App() {
       </div>
 
       <div className="row">
-        <div className="calendar-wrapper">
-          <BigCalendar
-            localizer={localizer}
-            events={events}
-            step={60}
-            defaultDate={currentDate}
-            popup={false}
-            onShowMore={(events, date) => this.setState({ showModal: true, events })}
-          />
-        </div>
+        <MaiselBaseCalendar currentDate={currentDate} events={events} />
       </div>
     </div>
   );
