+++ react-redux-client-app/src/App.js
@@ -20,7 +20,7 @@ function App() {
         </a>
       </header>
       <div>
-        <Counter/>
+        <Counter />
       </div>
     </div>
   );
